[Symbol, Traversal, Traversals/*] Change lookups to return `SymbolTable`s, and reorder the arguments to the `Ctx` `Child` constructor.
Returning `SymbolTable`s instead of `Symbol`s allows traversals to inspect child elements, which is useful e.g. when checking parameter types in a function call.
The arguments to `Child` used to suggest that `sym` was inbetween `lefts` and `rights`, where in reality it is the parent symbol. The constructor arguments have been reordered to remove this suggestion.